Add id to button and update JavaScript reference
In JavaScript, the button was referenced using it's class name,
'.register'. While this works just fine in a small project like this, a
best practice is to reference by id when an element is unique. Since
the button will only exist once in this app, we should reference it by
id.

Changes:

  - button now has an id='register'

  - JavaScript reference to button now uses id