Remove rounding from unit conversion functions
We need to ensure our functions only do what their name implies. In the
case of our conversion functions, we are only converting an input from
one unit of measurement to another. The rounding is not implicit in the
name of the function, so it should be performed outside the function.

Changes:

  - removed rounding from these conversion functions:

    - convertFahrenheitTo()

    - convertLbsTo()

  - inline rounding was added wherever the functions are called.